requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
jinja2
httpx
